+++ odoo/local-src/stock_product_bin/__manifest__.py
@@ -10,6 +10,7 @@
     'category': 'Stock',
     'depends': [
         'stock',
+        'stock_location_act_as_view',
     ],
     'data': [
         'views/product.xml',