+++ odoo/local-src/stock_receive_lot/wizards/stock.py
@@ -119,8 +119,9 @@ class StockPackOperationLotAdd(models.TransientModel):
 
     @api.onchange('operation_id')
     def _onchange_operation_id(self):
-        if self.operation_id.location_dest_id.usage == 'internal':
-            self.location_dest_id = self.operation_id.location_dest_id
+        op_dest_loc = self.operation_id.location_dest_id
+        if op_dest_loc.usage == 'internal' and not op_dest_loc.act_as_view:
+            self.location_dest_id = op_dest_loc
         elif not (
             self.location_dest_id
             and self._is_parent_child(
@@ -145,7 +146,7 @@ class StockPackOperationLotAdd(models.TransientModel):
     @api.depends('operation_id')
     def _get_location_op_dest_id(self):
         loc = self.operation_id.location_dest_id
-        while loc and loc.usage != 'view':
+        while loc and not loc.act_as_view:
             loc = loc.location_id
         self.location_op_dest_id = loc.id