+++ HISTORY.rst
@@ -23,6 +23,10 @@ latest (unreleased)
 
 **Bugfixes**
 
+* ALCYN-2043: change stock locations with usage 'view' to 'internal' with a new flag "act as view".
+  We cannot move quants to such location, but still they need to have an 'internal' usage to be considered
+  properly by the stock addon's code.
+
 **Build**
 
 **Documentation**